requirements.txt
torch
torchvision
albumentations
Pillow
numpy
pandas
matplotlib
seaborn
scikit-learn
PyYAML
tqdm
opencv-python